Move `\SMW\SQLStore\Table...` to `\SMW\SQLStore\TableBuilder`, refs 3801
